Use size_t instead of int

+++ strings.c
@@ -3,7 +3,7 @@
 
 char* strip_left(const char* input, const char delimiter) {
     int start = 0;
-    for (int i = 0; i < strlen(input); i++) {
+    for (size_t i = 0; i < strlen(input); i++) {
         if (input[i] != delimiter) {
             break;
         }
